updates to: ChartView,AdminUserView.
Made it so that the charts thread turns off if you turn off the program. It is now possible to see the number of hours next to the project name/user name in chartview. Made it so that if a project/user does not have hours it would not come up in charts.
When you click on a user in adminuserview, their name appears at the top of their taskview so it is easy to see for the admin which user's tasks he is currently looking at.